server/requirements.txt
connexion==1.0.103
django-dotenv==1.3.0
django-getenv==1.3.1